+++ ashen/src/asset/gamma_table.rs
@@ -66,6 +66,7 @@ impl GammaTable {
 }
 
 #[cfg(test)]
+#[cfg(feature = "conv")]
 mod tests {
     use std::cell::LazyCell;
