+++ ashen/src/asset/texture/dat/texture.rs
@@ -1,9 +1,11 @@
+use std::ops::Deref;
+
 use itertools::Itertools;
 
 use super::size::TextureSize;
 use crate::{
     asset::{Parser, color_map::Color},
-    utils::{format::PaletteTexture, nom::*},
+    utils::nom::*,
 };
 
 // TODO(nenikitov): Move this to a separate public module later
@@ -63,6 +65,24 @@ impl Parser for MippedTexture {
                 },
             ))
         }
+    }
+}
+
+pub trait PaletteTexture {
+    // TODO(Unavailable): `&[Color; 256]`
+    fn with_palette(&self, palette: &[Color]) -> Vec<Vec<Color>>;
+}
+
+// impl for any 2D array like data structure.
+impl<Outer: ?Sized, Inner> PaletteTexture for Outer
+where
+    Outer: Deref<Target = [Inner]>,
+    Inner: AsRef<[u8]>,
+{
+    fn with_palette(&self, palette: &[Color]) -> Vec<Vec<Color>> {
+        self.iter()
+            .map(|c| c.as_ref().iter().map(|c| palette[*c as usize]).collect())
+            .collect()
     }
 }
