+++ ashen/src/utils/mod.rs
@@ -1,5 +1,6 @@
 pub mod compression;
 // TODO(Unavailable): Rename to "conv"
+#[cfg(feature = "conv")]
 pub mod format;
 pub mod iterator;
 pub mod nom;
