+++ cli/Cargo.toml
@@ -4,4 +4,4 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
-ashen = { path = "../ashen/" }
+ashen = { path = "../ashen/", default-features = false }
